#21: Simplified make.py, added run_all.py, and started removing redundant traceback for Python 3
