+++ analysis/external.txt
@@ -15,3 +15,4 @@
 # This will create a file in `paths['external_dir']` named `destination` that is a copy of `source`.
 
 # Destination | Source
+wksmcldsk
