--- analysis/log/output_heads.log
@@ -1,28 +1,0 @@
-File headers
---------------------------------------------------------------------------------
-C:\Users\Zong\template\analysis\output\plot.eps
---------------------------------------------------------------------------------
-%!PS-Adobe-3.0 EPSF-3.0
-%%Title: output/plot.eps
-%%Creator: matplotlib version 2.0.2, http://matplotlib.org/
-%%CreationDate: Thu Jun  6 15:09:35 2019
-%%Orientation: portrait
-%%BoundingBox: 75 223 536 568
-%%EndComments
-%%BeginProlog
-/mpldict 8 dict def
-mpldict begin
---------------------------------------------------------------------------------
-C:\Users\Zong\template\analysis\output\tables.txt
---------------------------------------------------------------------------------
-<tab:table>
-300001.0
-173205.369
-600000
-2
-
-
-
-
-
---------------------------------------------------------------------------------
