--- analysis/log/output_stats.log
@@ -1,3 +1,0 @@
-file name | last modified | file size
-C:\Users\Zong\template\analysis\output\plot.eps | 2019-06-06 22:09:36 | 14084
-C:\Users\Zong\template\analysis\output\tables.txt | 2019-06-06 22:09:36 | 44
