--- analysis/log/source_heads.log
@@ -1,28 +1,0 @@
-File headers
---------------------------------------------------------------------------------
-C:\Users\Zong\template\data\output\data_graph.csv
---------------------------------------------------------------------------------
-x
-2
-4
-6
-8
-10
-12
-14
-16
-18
---------------------------------------------------------------------------------
-C:\Users\Zong\template\data\output\data_table.csv
---------------------------------------------------------------------------------
-"x"
-2
-4
-6
-8
-10
-12
-14
-16
-18
---------------------------------------------------------------------------------
