--- analysis/log/source_map.log
@@ -1,3 +1,0 @@
-destination | source
-input\data_table.csv | C:\Users\Zong\template\data\output\data_table.csv
-input\data_graph.csv | C:\Users\Zong\template\data\output\data_graph.csv
