--- analysis/log/source_stats.log
@@ -1,3 +1,0 @@
-file name | last modified | file size
-C:\Users\Zong\template\data\output\data_graph.csv | 2019-06-06 22:09:19 | 2344453
-C:\Users\Zong\template\data\output\data_table.csv | 2019-06-06 22:09:16 | 2344449
