+++ analysis/make.py
@@ -1,14 +1,18 @@
-### ENVIRONMENT
+###################
+### ENVIRONMENT ###
+###################
 import git
 import imp
 import os
 import yaml
 
+# LOAD GSLAB MAKE
 ROOT = git.Repo('.', search_parent_directories = True).working_tree_dir 
 f, path, desc = imp.find_module('gslab_make', [os.path.join(ROOT, 'lib')]) 
-gs = imp.load_module('gslab_make', f, path, desc) # Load `gslab_make`
+gs = imp.load_module('gslab_make', f, path, desc)
 
-PATHS = { # Set default paths; used by `gslab_make` functions
+# SET DEFAULT PATHS
+PATHS = {
     'config'          : '../config.yaml',
     'config_user'     : '../config_user.yaml',
     'input_dir'       : 'input', 
@@ -24,18 +28,20 @@ PATHS = { # Set default paths; used by `gslab_make` functions
     'source_headslog' : 'log/source_heads.log'  # Set to '' to avoid writing log
 }
 
-PATH_MAPPINGS = { # Set path mappings; used by `move_sources` functions
+### SET PATH MAPPINGS
+PATH_MAPPINGS = { 
     'root': ROOT
 }
 
-### CONFIG USER 
-config_user = yaml.load(open(PATHS['config_user'], 'rb'))
-if config_user['local']['executables']: # Update executables; used by `run_program` functions
-    gs.private.metadata.default_executables[os.name].update(config_user['local']['executables'])
-if config_user['external']: # Update path mappings; used by `create_links` functions
-    PATH_MAPPINGS.update(config_user['external'])
+### LOAD CONFIG USER 
+gs.update_executables(PATHS)
+PATH_MAPPINGS = gs.update_mappings(PATHS, PATH_MAPPINGS)
 
-### START 
+############
+### MAKE ###
+############
+
+### START MAKE
 gs.remove_dir(['input', 'external'])
 gs.clear_dir(['output', 'log'])
 gs.start_makelog(PATHS)
@@ -55,5 +61,5 @@ gs.log_files_in_output(PATHS)
 ### CHECK FILE SIZES
 gs.check_module_size(PATHS)
 
-### END
+### END MAKE
 gs.end_makelog(PATHS)
