--- analysis/output/tables.txt
@@ -1,5 +1,0 @@
-<tab:table>
-300001.0
-173205.369
-600000
-2
