+++ lib/gslab_make/__init__.py
@@ -40,10 +40,11 @@ Notes:
 from gslab_make.check_repo import check_module_size, get_modified_sources
 from gslab_make.move_sources import (create_links, link_inputs, link_externals, 
                                      create_copies, copy_inputs, copy_externals)
-from gslab_make.dir_mod import check_os, remove_path, remove_dir, clear_dir, unzip, zip_dir
+from gslab_make.modify_dir import remove_path, remove_dir, clear_dir, unzip, zip_dir
 from gslab_make.run_program import (run_stata, run_matlab, run_perl, run_python, 
                                     run_jupyter, run_mathematica, run_stat_transfer, 
                                     run_lyx, run_r, run_sas, execute_command)
+from gslab_make.update_env import check_os, update_executables, update_mappings
 from gslab_make.write_logs import (start_makelog, end_makelog, write_to_makelog,
                                    log_files_in_output, write_stats_log, write_heads_log)
 from gslab_make.write_source_logs import write_source_logs
