+++ paper_slides/input/plot.eps
@@ -1,7 +1,7 @@
 %!PS-Adobe-3.0 EPSF-3.0
 %%Title: output/plot.eps
 %%Creator: matplotlib version 2.0.2, http://matplotlib.org/
-%%CreationDate: Thu Jun  6 15:09:35 2019
+%%CreationDate: Fri Jul  5 20:29:51 2019
 %%Orientation: portrait
 %%BoundingBox: 75 223 536 568
 %%EndComments
