+++ paper_slides/log/make.log
@@ -1,5 +1,5 @@
 --------------------------------------------------------------------------------
-Makelog started: 2019-06-06 15:09:44
+Makelog started: 2019-07-05 20:29:54
 Working directory: C:\Users\Zong\template\paper_slides
 --------------------------------------------------------------------------------
 Input copies successfully created!
@@ -146,6 +146,6 @@ pdflatex: warning: running with administrator privileges
 
 Output logs successfully written!
 --------------------------------------------------------------------------------
-Makelog ended: 2019-06-06 15:10:03
+Makelog ended: 2019-07-05 20:30:18
 Working directory: C:\Users\Zong\template\paper_slides
 --------------------------------------------------------------------------------
