+++ paper_slides/log/output_heads.log
@@ -1,9 +1,5 @@
 File headers
 --------------------------------------------------------------------------------
-C:\Users\Zong\template\paper_slides\output\text.pdf
---------------------------------------------------------------------------------
-Head not readable or less than 10 lines
---------------------------------------------------------------------------------
 C:\Users\Zong\template\paper_slides\output\tables_filled.lyx
 --------------------------------------------------------------------------------
 #LyX 2.2 created this file. For more info see http://www.lyx.org/
@@ -17,7 +13,7 @@ C:\Users\Zong\template\paper_slides\output\tables_filled.lyx
 \maintain_unincluded_children false
 \language english
 --------------------------------------------------------------------------------
-C:\Users\Zong\template\paper_slides\output\ondeck.pdf
+C:\Users\Zong\template\paper_slides\output\text.pdf
 --------------------------------------------------------------------------------
 Head not readable or less than 10 lines
 --------------------------------------------------------------------------------
@@ -26,6 +22,10 @@ C:\Users\Zong\template\paper_slides\output\paper.pdf
 Head not readable or less than 10 lines
 --------------------------------------------------------------------------------
 C:\Users\Zong\template\paper_slides\output\online_appendix.pdf
+--------------------------------------------------------------------------------
+Head not readable or less than 10 lines
+--------------------------------------------------------------------------------
+C:\Users\Zong\template\paper_slides\output\ondeck.pdf
 --------------------------------------------------------------------------------
 Head not readable or less than 10 lines
 --------------------------------------------------------------------------------
