+++ paper_slides/log/output_stats.log
@@ -1,7 +1,7 @@
 file name | last modified | file size
-C:\Users\Zong\template\paper_slides\output\text.pdf | 2019-06-06 22:10:03 | 18008
-C:\Users\Zong\template\paper_slides\output\tables_filled.lyx | 2019-06-06 22:09:45 | 4338
-C:\Users\Zong\template\paper_slides\output\ondeck.pdf | 2019-06-06 22:10:01 | 18008
-C:\Users\Zong\template\paper_slides\output\paper.pdf | 2019-06-06 22:09:51 | 58309
-C:\Users\Zong\template\paper_slides\output\online_appendix.pdf | 2019-06-06 22:09:54 | 48135
-C:\Users\Zong\template\paper_slides\output\slides.pdf | 2019-06-06 22:09:59 | 38543
+C:\Users\Zong\template\paper_slides\output\tables_filled.lyx | 2019-07-06 03:29:55 | 4338
+C:\Users\Zong\template\paper_slides\output\text.pdf | 2019-07-06 03:30:18 | 18159
+C:\Users\Zong\template\paper_slides\output\paper.pdf | 2019-07-06 03:30:02 | 58502
+C:\Users\Zong\template\paper_slides\output\online_appendix.pdf | 2019-07-06 03:30:05 | 48333
+C:\Users\Zong\template\paper_slides\output\ondeck.pdf | 2019-07-06 03:30:14 | 18159
+C:\Users\Zong\template\paper_slides\output\slides.pdf | 2019-07-06 03:30:11 | 38543
