+++ paper_slides/output/ondeck.pdf
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:555542240f5cae3c82bad9f8eb3b1894f8d1317502462335b1cbb685cbc4311a
-size 18008
+oid sha256:698f53eef8d4e548799dca1424249fc4fc2d724ba2bb25e1e0fae8f4da5970f6
+size 18159
