+++ paper_slides/output/paper.pdf
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:839b3973a04b9c98667d29b4a34ad5ae91575b9aa2f125c507da54416d892d95
-size 58309
+oid sha256:40f272f5df468b1de643e285fcb826d7224e626cb5819878ca19853bb7a2d360
+size 58502
