+++ paper_slides/output/text.pdf
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:908b0152f41d67a68c3980264cb2449823d6f6fc5ae1f729e437c15bacbc820e
-size 18008
+oid sha256:f9587fc19c00dafc9d74a3965bcad43f5ed4cfb83d36e9c4318f706a24c81ecf
+size 18159
